Prevent Emailing Duplicate Errors
Added a new PreventDuplicates configuration option to the ErrorEmailer
with the related code changes necessary to prevent the ErrorEmailer from
mailing notifications when the error was detected to be a duplicate by
the ErrorStore as configured by the RollupThreshold value. One of those
changes was the addition of a simple bool flag on Error called
IsOriginalError, which should only be true when the error was not
determined to be a duplicate of a previously stored value, or if it was
queued and the queue didn't already contain a matching error.