Added LastDuplicateDate
In agreement with the issue NickCraver/StackExchange.Exceptional#31,
I've added LastDuplicateDate to the Error class and added code for
conditionally (based on config options) using it in the ErrorStore and
related classes.

Making it configurable means that users who aren't able to change their
storage tables would still be able to use Exceptional in exactly the
same way as before. Adding functionality without losing compatibility
makes it a win-win. ;)